% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{bank}
\format{
\subsection{\code{bank}}{

A data frame with 100 rows and 3 columns:
\describe{
\item{arrival_time}{time they arrive}
\item{service_time}{time for servie}
...
}
}
}
\usage{
bank
}
\description{
Dataset used to simulate customer queuing.
}
\keyword{datasets}
