% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Simulating a customer queuing system}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{A positive numeric value for inter arrival times}

\item{ServiceTimes}{A positive numeric value for service times}

\item{NumServers}{An integer which dictates the number of servers in the queuing system.}
}
\value{
Will return as tibble:tibble() containing information about the queuing system.
}
\description{
Simulating a customer queuing system
}
\examples{
data(bank)
with(bank, Multiserver(arrival_time, service_time, 6))
}
